% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_plot.R
\name{histogram_plot}
\alias{histogram_plot}
\title{Use geom_histogram() to make a plot}
\usage{
histogram_plot(data, ylab_str, his_width)
}
\arguments{
\item{infile}{path to the input file; the ylab title; width for his}
}
\value{
a broken line graph for your input file
}
\description{
This function loads a file as a matrix and makes a plot for it.The file must be whithout a header.
}
