% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plot.R
\name{line_plot}
\alias{line_plot}
\title{Use geom_line() to make a plot}
\usage{
line_plot(data, ylab_str, ymin, ymax)
}
\arguments{
\item{infile}{path to the input file; the ylab title; ymin of ylim; ymax of ylim}
}
\value{
a broken line graph for your input file
}
\description{
This function loads a file as a matrix and makes a plot for it.The file must be whithout a header.
}
