% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines_plot.R
\name{lines_plot}
\alias{lines_plot}
\title{Use geom_line() to make a plot}
\usage{
lines_plot(data, ylab_str, order_type)
}
\arguments{
\item{infile}{path to the input file; the ylab title; the type name used for ordering}
}
\value{
a broken line graph for your input file
}
\description{
This function loads a file as a matrix and makes a plot for it.The file must be whithout a header.
}
