% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneline_histogram_plot.R
\name{oneline_histogram_plot}
\alias{oneline_histogram_plot}
\title{Use geom_line() and geom_histogram() to make a plot}
\usage{
oneline_histogram_plot(
  data,
  ylab_str,
  line_type,
  histogram_type,
  order_type,
  ymin,
  ymax
)
}
\arguments{
\item{infile}{path to the input file; the ylab title; the type name used in geom_line(); the type name used in geom_histogram(); the type name used for orderring ; ymin of ylim; ymax of ylim}
}
\value{
a broken line graph for your input file
}
\description{
This function loads a file as a matrix and makes a plot for it.The file must be whithout a header.
}
