% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sv_count.R
\name{sv_count}
\alias{sv_count}
\title{Use geom_line() && geom_text_repel() && geom_point() to make a plot}
\usage{
sv_count(data, ylab_str)
}
\arguments{
\item{infile}{path to the input file; the ylab title;}
}
\value{
a plot for your input file
}
\description{
This function loads a file as a matrix and makes a plot for it.The file must be whithout a header.
}
